+++ email/src/imap/config.rs
@@ -3,7 +3,9 @@
 //! This module contains the implementation of the IMAP backend and
 //! all associated structures related to it.
 
-use serde::{Deserialize, Serialize};
+use imap::ConnectionMode;
+use serde::{de, Deserialize, Deserializer, Serialize};
+use std::{fmt, marker::PhantomData, result};
 use thiserror::Error;
 
 use crate::{
@@ -14,8 +16,6 @@ use crate::{
 /// Errors related to the IMAP backend configuration.
 #[derive(Debug, Error)]
 pub enum Error {
-    #[error("cannot start the notify mode")]
-    StartNotifyModeError(#[source] process::Error),
     #[error("cannot get imap password from global keyring")]
     GetPasswdError(#[source] secret::Error),
     #[error("cannot get imap password: password is empty")]
@@ -32,20 +32,11 @@ pub struct ImapConfig {
     /// The IMAP server host port.
     pub port: u16,
 
-    /// Enables TLS/SSL.
+    /// The IMAP encryption protocol to use.
     ///
-    /// Defaults to `true`.
-    pub ssl: Option<bool>,
-
-    /// Enables StartTLS.
-    ///
-    /// Defaults to `false`.
-    pub starttls: Option<bool>,
-
-    /// Trusts any certificate.
-    ///
-    /// Defaults to `false`.
-    pub insecure: Option<bool>,
+    /// Supported encryption: SSL/TLS, STARTTLS or none.
+    #[serde(default, deserialize_with = "some_bool_or_kind")]
+    pub encryption: Option<ImapEncryptionKind>,
 
     /// The IMAP server login.
     ///
@@ -68,19 +59,30 @@ pub struct ImapConfig {
 }
 
 impl ImapConfig {
-    /// TLS/SSL option getter.
-    pub fn ssl(&self) -> bool {
-        self.ssl.unwrap_or(true)
+    /// Return `true` if TLS or StartTLS is enabled.
+    pub fn is_encryption_enabled(&self) -> bool {
+        match self.encryption.as_ref() {
+            None => true,
+            Some(ImapEncryptionKind::Tls) => true,
+            Some(ImapEncryptionKind::StartTls) => true,
+            _ => false,
+        }
     }
 
-    /// StartTLS option getter.
-    pub fn starttls(&self) -> bool {
-        self.starttls.unwrap_or_default()
+    /// Return `true` if StartTLS is enabled.
+    pub fn is_start_tls_encryption_enabled(&self) -> bool {
+        match self.encryption.as_ref() {
+            Some(ImapEncryptionKind::StartTls) => true,
+            _ => false,
+        }
     }
 
-    /// Insecure option getter.
-    pub fn insecure(&self) -> bool {
-        self.insecure.unwrap_or_default()
+    /// Return `true` if encryption is disabled.
+    pub fn is_encryption_disabled(&self) -> bool {
+        match self.encryption.as_ref() {
+            Some(ImapEncryptionKind::None) => true,
+            _ => false,
+        }
     }
 
     /// Builds authentication credentials.
@@ -94,6 +96,47 @@ impl ImapConfig {
     /// Find the IMAP watch timeout.
     pub fn find_watch_timeout(&self) -> Option<u64> {
         self.watch.as_ref().and_then(|c| c.find_timeout())
+    }
+}
+
+#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "kebab-case")]
+pub enum ImapEncryptionKind {
+    #[default]
+    #[serde(alias = "ssl")]
+    Tls,
+    #[serde(alias = "starttls")]
+    StartTls,
+    None,
+}
+
+impl fmt::Display for ImapEncryptionKind {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Tls => write!(f, "SSL/TLS"),
+            Self::StartTls => write!(f, "StartTLS"),
+            Self::None => write!(f, "None"),
+        }
+    }
+}
+
+impl From<bool> for ImapEncryptionKind {
+    fn from(value: bool) -> Self {
+        if value {
+            Self::Tls
+        } else {
+            Self::None
+        }
+    }
+}
+
+impl Into<ConnectionMode> for ImapEncryptionKind {
+    fn into(self) -> ConnectionMode {
+        match self {
+            Self::Tls => ConnectionMode::Tls,
+            Self::StartTls => ConnectionMode::StartTls,
+            Self::None => ConnectionMode::Plaintext,
+        }
     }
 }
 
@@ -175,4 +218,56 @@ impl ImapWatchConfig {
     pub fn find_timeout(&self) -> Option<u64> {
         self.timeout
     }
+}
+
+fn some_bool_or_kind<'de, D>(
+    deserializer: D,
+) -> result::Result<Option<ImapEncryptionKind>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    struct SomeBoolOrKind(PhantomData<fn() -> Option<ImapEncryptionKind>>);
+
+    impl<'de> de::Visitor<'de> for SomeBoolOrKind {
+        type Value = Option<ImapEncryptionKind>;
+
+        fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+            formatter.write_str("some or none")
+        }
+
+        fn visit_some<D>(self, deserializer: D) -> result::Result<Self::Value, D::Error>
+        where
+            D: Deserializer<'de>,
+        {
+            struct BoolOrKind(PhantomData<fn() -> ImapEncryptionKind>);
+
+            impl<'de> de::Visitor<'de> for BoolOrKind {
+                type Value = ImapEncryptionKind;
+
+                fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+                    formatter.write_str("boolean or string")
+                }
+
+                fn visit_bool<E>(self, v: bool) -> result::Result<Self::Value, E>
+                where
+                    E: de::Error,
+                {
+                    Ok(v.into())
+                }
+
+                fn visit_str<E>(self, v: &str) -> result::Result<Self::Value, E>
+                where
+                    E: de::Error,
+                {
+                    Deserialize::deserialize(de::value::StrDeserializer::new(v))
+                }
+            }
+
+            deserializer
+                .deserialize_any(BoolOrKind(PhantomData))
+                .map(Option::Some)
+        }
+    }
+
+    deserializer.deserialize_option(SomeBoolOrKind(PhantomData))
 }