+++ email/src/smtp/config.rs
@@ -5,8 +5,8 @@
 
 use log::debug;
 use mail_send::Credentials;
-use serde::{Deserialize, Serialize};
-use std::io;
+use serde::{de, Deserialize, Deserializer, Serialize};
+use std::{fmt, io, marker::PhantomData, result};
 use thiserror::Error;
 
 use crate::{
@@ -35,20 +35,11 @@ pub struct SmtpConfig {
     /// The SMTP server host port.
     pub port: u16,
 
-    /// Enables TLS/SSL.
+    /// The SMTP encryption protocol to use.
     ///
-    /// Defaults to `true`.
-    pub ssl: Option<bool>,
-
-    /// Enables StartTLS.
-    ///
-    /// Defaults to `false`.
-    pub starttls: Option<bool>,
-
-    /// Trusts any certificate.
-    ///
-    /// Defaults to `false`.
-    pub insecure: Option<bool>,
+    /// Supported encryption: SSL/TLS or STARTTLS.
+    #[serde(default, deserialize_with = "some_bool_or_kind")]
+    pub encryption: Option<SmtpEncryptionKind>,
 
     /// The SMTP server login.
     ///
@@ -65,6 +56,32 @@ pub struct SmtpConfig {
 }
 
 impl SmtpConfig {
+    /// Return `true` if TLS or StartTLS is enabled.
+    pub fn is_encryption_enabled(&self) -> bool {
+        match self.encryption.as_ref() {
+            None => true,
+            Some(SmtpEncryptionKind::Tls) => true,
+            Some(SmtpEncryptionKind::StartTls) => true,
+            _ => false,
+        }
+    }
+
+    /// Return `true` if StartTLS is enabled.
+    pub fn is_start_tls_encryption_enabled(&self) -> bool {
+        match self.encryption.as_ref() {
+            Some(SmtpEncryptionKind::StartTls) => true,
+            _ => false,
+        }
+    }
+
+    /// Return `true` if encryption is disabled.
+    pub fn is_encryption_disabled(&self) -> bool {
+        match self.encryption.as_ref() {
+            Some(SmtpEncryptionKind::None) => true,
+            _ => false,
+        }
+    }
+
     /// Builds the SMTP credentials string.
     ///
     /// The result depends on the [`SmtpAuthConfig`]: if password mode
@@ -85,20 +102,36 @@ impl SmtpConfig {
             },
         })
     }
+}
 
-    /// SSL enabled getter.
-    pub fn ssl(&self) -> bool {
-        self.ssl.unwrap_or(true)
+#[derive(Clone, Debug, Default, Eq, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "kebab-case")]
+pub enum SmtpEncryptionKind {
+    #[default]
+    #[serde(alias = "ssl")]
+    Tls,
+    #[serde(alias = "starttls")]
+    StartTls,
+    None,
+}
+
+impl fmt::Display for SmtpEncryptionKind {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Tls => write!(f, "SSL/TLS"),
+            Self::StartTls => write!(f, "StartTLS"),
+            Self::None => write!(f, "None"),
+        }
     }
+}
 
-    /// STARTTLS enabled getter.
-    pub fn starttls(&self) -> bool {
-        self.starttls.unwrap_or_default()
-    }
-
-    /// Insecure mode getter
-    pub fn insecure(&self) -> bool {
-        self.insecure.unwrap_or_default()
+impl From<bool> for SmtpEncryptionKind {
+    fn from(value: bool) -> Self {
+        if value {
+            Self::Tls
+        } else {
+            Self::None
+        }
     }
 }
 
@@ -167,4 +200,56 @@ impl SmtpAuthConfig {
             }
         }
     }
+}
+
+fn some_bool_or_kind<'de, D>(
+    deserializer: D,
+) -> result::Result<Option<SmtpEncryptionKind>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    struct SomeBoolOrKind(PhantomData<fn() -> Option<SmtpEncryptionKind>>);
+
+    impl<'de> de::Visitor<'de> for SomeBoolOrKind {
+        type Value = Option<SmtpEncryptionKind>;
+
+        fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+            formatter.write_str("some or none")
+        }
+
+        fn visit_some<D>(self, deserializer: D) -> result::Result<Self::Value, D::Error>
+        where
+            D: Deserializer<'de>,
+        {
+            struct BoolOrKind(PhantomData<fn() -> SmtpEncryptionKind>);
+
+            impl<'de> de::Visitor<'de> for BoolOrKind {
+                type Value = SmtpEncryptionKind;
+
+                fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+                    formatter.write_str("boolean or string")
+                }
+
+                fn visit_bool<E>(self, v: bool) -> result::Result<Self::Value, E>
+                where
+                    E: de::Error,
+                {
+                    Ok(v.into())
+                }
+
+                fn visit_str<E>(self, v: &str) -> result::Result<Self::Value, E>
+                where
+                    E: de::Error,
+                {
+                    Deserialize::deserialize(de::value::StrDeserializer::new(v))
+                }
+            }
+
+            deserializer
+                .deserialize_any(BoolOrKind(PhantomData))
+                .map(Option::Some)
+        }
+    }
+
+    deserializer.deserialize_option(SomeBoolOrKind(PhantomData))
 }