+++ email/tests/sendmail.rs
@@ -7,7 +7,7 @@ async fn test_sendmail_features() {
         envelope::list::imap::ListEnvelopesImap,
         folder::purge::imap::PurgeFolderImap,
         imap::{
-            config::{ImapAuthConfig, ImapConfig},
+            config::{ImapAuthConfig, ImapConfig, ImapEncryptionKind},
             ImapSessionBuilder,
         },
         message::send_raw::sendmail::SendRawMessageSendmail,
@@ -23,9 +23,7 @@ async fn test_sendmail_features() {
     let imap_config = ImapConfig {
         host: "localhost".into(),
         port: 3143,
-        ssl: Some(false),
-        starttls: Some(false),
-        insecure: Some(true),
+        encryption: Some(ImapEncryptionKind::None),
         login: "bob@localhost".into(),
         auth: ImapAuthConfig::Passwd(PasswdConfig {
             passwd: Secret::new_raw("password"),